STL of fancier rubber scroll wheel bit
Printed In TPU at .05 layer height 100% infil, or enough walls to make it 100%. I used the geometry/faces from the provided scroll wheel, which is CERN licensed, or else it would go on my printables page. This seems like the next best option to share my files. the .blend file will be uploaded with the step files in this fork in case you want to modify it further :-)